backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.31.0
transformers>=4.30.0
torch>=2.0.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
sentence-transformers>=2.2.0
